Bug Fix: allow emitted particles to live out their lives on stopEmit() (#208)
* Bug Fix: allow emitted particles to live out their lives on stopEmit()

* Bug: handle last particle lingering on death

* Update sandbox/common/Visualization.js

Co-authored-by: Rohan Deshpande <rohan@creativelifeform.com>

* Fixed failing test

So this test was failing, because with these changes, if there is no emitter.particles.length then we will never call this.dispatch(SYSTEM_UPDATE) and in this test we are simply creating the emitter but not starting it so it will never have any particles.  So the fix in this situation is to simply expect two calls instead of three.

A better test would be to actually start/stop the emitter to emit some particles and verify that they live out their full lives.  Im sure this is possible, but it feels like a complex test.

* Added particle life test

Co-authored-by: Andrew Leek <andrewleek@gmail.com>
Co-authored-by: Rohan Deshpande <rohan@creativelifeform.com>