Particle rotation and graceful emitter EOL (#222) (#223)
* Bug Fix: allow emitted particles to live out their lives on stopEmit() (#208)

* Bug Fix: allow emitted particles to live out their lives on stopEmit()

* Bug: handle last particle lingering on death

* Update sandbox/common/Visualization.js

Co-authored-by: Rohan Deshpande <rohan@creativelifeform.com>

* Fixed failing test

So this test was failing, because with these changes, if there is no emitter.particles.length then we will never call this.dispatch(SYSTEM_UPDATE) and in this test we are simply creating the emitter but not starting it so it will never have any particles.  So the fix in this situation is to simply expect two calls instead of three.

A better test would be to actually start/stop the emitter to emit some particles and verify that they live out their full lives.  Im sure this is possible, but it feels like a complex test.

* Added particle life test

Co-authored-by: Andrew Leek <andrewleek@gmail.com>
Co-authored-by: Rohan Deshpande <rohan@creativelifeform.com>

* Enhancement - enabled particle rotations (#210)

* enhancement: enabled particle rotation

- added support for both SpriteRenderer & GPURenderer

* removed unnecessary eulerFromDir step

- after looking into the internals of the three.js Euler class I found that it is simply copying x,y,z from the vector3, so we can just do the same with vector3.copy. (this also removes the unnecessary instantiation of a new Euler)

* additionally I think it would be nice to eventually remove Vector3D and simply use three.js’s Vector3 class, removing this gets us closer to that goal.  This may simply be an opinion, but I feel it would be nice to keep things as close to the standard three.js types as possible.  Perhaps this could be the subject of further discussion.

* Added Rotation Initializer

* Initializer:Rotation - option so make rotation emitter/parent independent

* Fixed Sprite set rotation test

This was a check to protect against setting the rotation if the target was a THREE Sprite, but now that this functionality has been added, it checks FOR setting rotation on a THREE Sprite.  Since a THREE Sprite can now handle rotation correctly, perhaps this check should simply be removed as it should be covered by the following test: 'should copy the particle position, set its rotation and scale it if it has a target'

Co-authored-by: Andrew Leek <andrewleek@gmail.com>

* Did a quick smoke test to make sure all okay, seems good

* Fixed vulns

Co-authored-by: Andrew Leek <64859328+andrewleek-droplab@users.noreply.github.com>
Co-authored-by: Andrew Leek <andrewleek@gmail.com>

Co-authored-by: Andrew Leek <64859328+andrewleek-droplab@users.noreply.github.com>
Co-authored-by: Andrew Leek <andrewleek@gmail.com>